% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lj.z.test.R
\name{lj.z.test}
\alias{lj.z.test}
\title{Corrected Z-test of Looney and Jones}
\usage{
lj.z.test(x, y, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
p.value /---/ p value of the test

statistic /---/ the value of the z-statistic.
}
\description{
Perform two sample z test of Looney and Jones on vectors of data.
}
\details{
--------------Needs to be filled...------------------
}
\examples{
vec1 <- c(2,3,NA,4,3,4,NA,5,4,2)
vec2 <- c(3,NA,4,4,3,NA,NA,4,NA,2)
lj.z.test(vec1, vec2, alternative = "greater")

}
