% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.z.test.R
\name{weighted.z.test}
\alias{weighted.z.test}
\title{Weighted Z-test combination}
\usage{
weighted.z.test(x, y, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
p.valuethe /---/ p value of the test

statistic /---/ the value of the t-statistic.
}
\description{
Perform two sample weighted z test on vectors of data.
}
\details{
--------------Needs to be filled...------------------
}
\examples{
vec1 <- c(2,3,NA,4,3,4,NA,5,4,2)
vec2 <- c(3,NA,4,4,3,NA,NA,4,NA,2)
weighted.z.test(vec1, vec2, alternative = "greater")

}
